client: Use sane default values for the custom transport
When setting up the custom transport for TLS support, the values for
timeouts, keepalive, and connection counts are all defaulting to zero
values. This can cause the client to not reconnect to a server that has
become unavailable.

Use the values of DefaultTransport in net/http to set up the custom
transport with appropriate settings.